refactor(MenuOption): 패키지 위치 변경과 분기 관련 메서드 및 객체 삭제
